requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
flask-socketio==5.1.1
python-engineio==4.2.1
python-socketio==5.4.0